my-personal-site/backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
pydantic[email]
mysql-connector-python
python-dotenv
bcrypt
pyjwt
passlib
